Renamed eqx.tree_inference -> eqx.nn.inference_mode, as it's really an eqx.nn thing, not an eqx thing.
