+++ docs/api/nn/inference.md
@@ -1,0 +1,3 @@
+# Training/Inference
+
+::: equinox.nn.inference_mode
