+++ equinox/nn/_dropout.py
@@ -16,7 +16,7 @@ class Dropout(Module):
     Note that this layer behaves differently during training and inference. During
     training then dropout is randomly applied; during inference this layer does nothing.
     Whether the model is in training or inference mode should be toggled using
-    [`equinox.tree_inference`][].
+    [`equinox.nn.inference_mode`][].
     """
 
     # Not static fields as it makes sense to want to modify them via equinox.tree_at.
@@ -35,7 +35,7 @@ class Dropout(Module):
         - `p`: The fraction of entries to set to zero. (On average.)
         - `inference`: Whether to actually apply dropout at all. If `True` then dropout
             is *not* applied. If `False` then dropout is applied. This may be toggled
-            with [`equinox.tree_inference`][] or overridden during
+            with [`equinox.nn.inference_mode`][] or overridden during
             [`equinox.nn.Dropout.__call__`][].
         - `deterministic`: Deprecated alternative to `inference`.
         """
